adjust sqlite imports/exports
